requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appdirs==1.4.3
apturl==0.5.2
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
base58==2.1.1
bcrypt==3.1.7
bitarray==1.2.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.12
Click==7.0
cloudinary==1.28.1
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
cytoolz==0.11.2
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-phonenumber-field==6.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.3.0
eth-typing==2.3.0
eth-utils==1.10.0
fasteners==0.14.1
filelock==3.0.12
frozenlist==1.3.0
future==0.18.2
greenlet==0.4.15
hexbytes==0.2.2
httplib2==0.14.0
idna==2.8
importlib-metadata==4.10.1
importlib-resources==5.4.0
inflection==0.5.1
ipfshttpclient==0.8.0a2
itypes==1.2.0
Jinja2==3.0.3
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lru-dict==1.1.7
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
msgpack==0.6.2
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.6.0
parsimonious==0.8.1
pexpect==4.6.0
phonenumberslite==8.12.41
Pillow==7.0.0
protobuf==3.19.4
psycopg2==2.8.4
pycairo==1.16.2
pycodestyle==2.8.0
pycryptodome==3.14.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pynvim==0.4.1
pyparsing==3.0.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyTelegramBotAPI==4.4.0
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.6
python-dotenv==0.19.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rlp==2.0.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
swagger-spec-validator==2.7.4
systemd-python==234
toml==0.10.2
toolz==0.11.2
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
varint==1.0.2
virtualenv==20.0.17
wadllib==1.3.3
web3==5.28.0
websockets==9.1
xkit==0.0.0
yarl==1.7.2
zipp==3.7.0
